+++ code/hsec-tools/CHANGELOG.md
@@ -4,6 +4,9 @@
 * Add support for GHC component in `query is-affected`
 * Add `model.database_specific.{repository,osvs,home}` and `model.affected.database_specific.{osv,human_link}` in OSV exports
 * Adapt to new security-advisories layout
+* Drop `Security.Advisories.Filesystem.parseComponentIdentifier`
+* Drop `Security.Advisories.Parse.OutOfBandAttributes.oobComponentIdentifier`
+* Drop `Security.Advisories.Parse.OOBError.PathHasNoComponentIdentifier`
 
 ## 0.2.0.2