+++ code/hsec-tools/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0.1
+
+* Bump `hsec-core` `0.3.0.0`
+
 ## 0.3.0.0
 
 * Move `isVersionAffectedBy` and `isVersionRangeAffectedBy` to `Security.Advisories.Core` (`hsec-core`)